Scenario

Genesys.Scenarios — Type

function Scenarios(mg::Microgrid, d::Dict{})

Constructor function for creating a new Scenarios instance based on a given Microgrid and a Dict containing scenario data.

Arguments

mg::Microgrid: A Microgrid instance.

d::Dict{}: A dictionary containing scenario data.

Returns

Scenarios: A Scenarios instance with the specified data.

Example

microgrid = ...
scenario_data = ...

scenarios = Scenarios(microgrid, scenario_data)

Scenarios Reducer

Genesys.ManualReducer — Type

mutable struct ManualReducer <: AbstractScenariosReducer
    h::Union{UnitRange{Int64}, Int64}
    y::Union{UnitRange{Int64}, Int64}
    s::Union{UnitRange{Int64}, Int64}
end

A mutable structure representing a manual scenario reduction. Inherits from the AbstractScenariosReducer abstract type.

Fields

h::Union{UnitRange{Int64}, Int64}: The hours to be included in the reduced scenario.

y::Union{UnitRange{Int64}, Int64}: The years to be included in the reduced scenario.

s::Union{UnitRange{Int64}, Int64}: The scenarios to be included in the reduced scenario.

Constructor

ManualReducer(; h = 1:8760, y = 2:3, s = 1:1) = new(h, y, s)

Genesys.SAAReducer — Type

mutable struct SAAReducer <: AbstractScenariosReducer
    nsample::Int64
end

A mutable structure representing a Sample Average Approximation (SAA) scenario reduction. Inherits from the AbstractScenariosReducer abstract type.

Fields

nsample::Int64: The number of samples to be included in the reduced scenario.

Constructor

SAAReducer(; nsample = 100) = new(nsample)

Genesys.MeanValueReducer — Type

mutable struct MeanValueReducer <: AbstractScenariosReducer
end

A mutable structure representing an Expected Value (Mean Value) scenario reduction. Inherits from the AbstractScenariosReducer abstract type.

Constructor

MeanValueReducer() = new()

Genesys.FeatureBasedReducer — Type

mutable struct FeatureBasedReducer <: AbstractScenariosReducer
    transformation::Union{UnionAll, Nothing}
    reduction::Union{AbstractDimensionReducer, Nothing}
    clustering::AbstractClusteringMethod
end

A mutable structure representing a Clustering scenario reduction using feature-based methods. Inherits from the AbstractScenariosReducer abstract type.

Fields

transformation: A transformation method to apply before clustering (default: UnitRangeTransform)

reduction: A dimensionality reduction method to apply before clustering (default: StatsReduction())

clustering: The clustering method to use (default: KmedoidsClustering())

Constructor

FeatureBasedReducer(; transformation = UnitRangeTransform,
                    reduction = StatsReduction(),
                    clustering = KmedoidsClustering()) = new(transformation, reduction, clustering)

Dimention Reducer

Genesys.PCAReduction — Type

mutable struct PCAReduction <: AbstractDimensionReducer
    n_components::Int64
end

A mutable structure representing Principal Component Analysis (PCA) dimensionality reduction. Inherits from the AbstractDimensionReducer abstract type.

Fields

n_components: The number of principal components to use for dimensionality reduction (default: 2)

Constructor

PCAReduction(; n_components = 2) = new(n_components)

Genesys.StatsReduction — Type

struct StatsReduction <: AbstractDimensionReducer

A structure representing statistics-based dimensionality reduction. Inherits from the AbstractDimensionReducer abstract type.

Constructor

StatsReduction()

Transformation

StatsBase.UnitRangeTransform — Type

Unit range normalization

StatsBase.ZScoreTransform — Type

Standardization (Z-score transformation)

Custering

Genesys.KmedoidsClustering — Type

mutable struct KmedoidsClustering <: AbstractClusteringMethod

A structure representing K-medoids clustering method. Inherits from the AbstractClusteringMethod abstract type.

Fields

n_clusters::Int64: Number of clusters to form.

distance::Distances.SemiMetric: Distance metric to use for calculating the dissimilarity between points.

log::Bool: Whether to print logs during the clustering process.

Constructor

KmedoidsClustering(; n_clusters = 20, distance = Distances.Euclidean(), log = true)

Generator

Genesys.MarkovGenerator — Type

mutable struct MarkovGenerator <: AbstractScenariosGenerator

A structure representing a Markov chain-based scenario generator.

Fields

nstate::Int64: Number of states, which corresponds to the number of clusters per hour per month.

algo::String: Algorithm used for clustering the states.

markovchains::NamedTuple{(:wk, :wkd), Tuple{MarkovChain, MarkovChain}}: A named tuple containing the Markov chains for weekdays and weekends.

Example

markov_generator = MarkovGenerator(nstate=10, algo="kmeans")

Genesys.AnticipativeGenerator — Type

mutable struct AnticipativeGenerator <: AbstractScenariosGenerator

A structure representing an anticipative scenario generator (perfect foresight).

Fields

forecast: The field to store perfect foresight information.

Example

anticipative_generator = AnticipativeGenerator()

Base generic function

Generate

Genesys.generate — Function

Generate scenarios using a Markov chain.

Parameters: generator: a MarkovGenerator object s0: a ny x generator.nstate matrix representing the initial state t0: a DateTime object representing the initial timestamp nstep: an integer representing the number of steps to generate ny: an integer representing the number of scenarios to generate ns: an integer representing the number of seeds for the generator

Returns:

a list of ny ny x ns x nstep matrix representing the generated scenarios

a ny x ns matrix representing the probability associated with each generated scenario

generate(generator::AnticipativeGenerator, s0, t0::DateTime, nstep::Int64; ny::Int64=1, ns::Int64=1, h::Int64=1)

Generate perfect forecast.

Arguments

generator::AnticipativeGenerator: Anticipative generator struct.

s0: Vector of initial state(s).

t0::DateTime: Initial time.

nstep::Int64: Number of steps to simulate.

ny::Int64=1: Number of outputs.

ns::Int64=1: Number of scenarios.

h::Int64=1: Initial time step.

Returns

forecast: Matrix of shape (nstep, ny, ns) with the forecasted values.

probs: Array of shape (ny, ns) with the probability distribution of each scenario.

Reduce

Genesys.reduce — Function

function reduce(reducer::ManualReducer, ω::Scenarios)

Reduces the given Scenarios object using the specified ManualReducer.

Arguments

reducer::ManualReducer: The reducer used to specify the reduction process.

ω::Scenarios: The Scenarios object to be reduced.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario.

Example

manual_reducer = ManualReducer(h = 1:8760, y = 2:3, s = 1:1)
reduced_scenarios, probabilities = reduce(manual_reducer, scenarios)

function reduce(reducer::SAAReducer, ω::Scenarios; y::Int64 = 1, s::Int64 = 1)

Reduces the given Scenarios object using the specified SAAReducer, implementing the Sample Average Approximation scenario reduction method.

Arguments

reducer::SAAReducer: The reducer used to specify the reduction process.

ω::Scenarios: The Scenarios object to be reduced.

Keyword Arguments

y::Int64 = 1: The initial year to be considered.

s::Int64 = 1: The initial scenario to be considered.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario.

Example

saa_reducer = SAAReducer(nsample = 100)
reduced_scenarios, probabilities = reduce(saa_reducer, scenarios)

function reduce(reducer::MeanValueReducer, ω::Scenarios; y::Int64 = 1, s::Int64 = 1)

Reduces the given Scenarios object using the specified MeanValueReducer, implementing the Expected Value scenario reduction method.

Arguments

reducer::MeanValueReducer: The reducer used to specify the reduction process.

ω::Scenarios: The Scenarios object to be reduced.

Keyword Arguments

y::Int64 = 1: The initial year to be considered.

s::Int64 = 1: The initial scenario to be considered.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario (in this case, a single element array containing 1).

Example

mean_value_reducer = MeanValueReducer()
reduced_scenarios, probabilities = reduce(mean_value_reducer, scenarios)

function reduce(reducer::FeatureBasedReducer, ω::Scenarios; y::Int64 = 1, s::Int64 = 1)

Reduces the given Scenarios object using the specified FeatureBasedReducer, implementing a clustering-based scenario reduction method.

Arguments

reducer::FeatureBasedReducer: The reducer used to specify the reduction process, including transformation, dimension reduction, and clustering methods.

ω::Scenarios: The Scenarios object to be reduced.

Keyword Arguments

y::Int64 = 1: The initial year to be considered.

s::Int64 = 1: The initial scenario to be considered.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario in the reduced set.

An array of assignments indicating which cluster each original scenario belongs to.

Example

feature_based_reducer = FeatureBasedReducer(transformation = UnitRangeTransform,
                                            reduction = StatsReduction(),
                                            clustering = KmedoidsClustering())
reduced_scenarios, probabilities, assignments = reduce(feature_based_reducer, scenarios)