Assets

Storages

Batteries

Main.Genesys.AbstractLiion — Type

abstract type AbstractLiion <: AbstractStorage

An abstract type to generalize every Liion battery storage solution

Main.Genesys.Liion_rainflow — Type

Liion_rainflow

A mutable struct that represents a Li-ion battery model with rainflow counting algorithm for State of Health (SoH) computation.

The structure have a lot of parameters but most of them are set to default values.

Parameters:

α_p_ch::Float64: Charging maximum C-rate (default : 1.5)

α_p_dch::Float64: Discharging maximum C-rate (default : 1.5)

α_soc_min::Float64: Minimum threshold of charge (normalized) (default : 0.2)

α_soc_max::Float64: Maximum threshold of charge (normalized) (default : 0.8)

SoH_threshold::Float64: SoH level to replace the battery (default : 0.8)

couplage::NamedTuple: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling)

soc_model::String: Model name for State of Charge (SoC) computation. Available models are listed

calendar::Bool: Whether to include calendar aging in the SoH computation (default : true)

soc_ini::Float64: Initial State of Charge (SoC) for the beginning of the simulation (default : 0.5)

soh_ini::Float64: Initial State of Health (SoH) for the beginning of the simulation (default : 1)

update_by_year::Int64: Rainflow SoH computation by year (default : 12)

fatigue_data::DataFrames.DataFrame: DataFrame containing fatigue data (DoD, ncycle) (Default NMC battery data are provided)

Example of usage : Liion_rainflow(update_by_year = 12, calendar = true, soc_model = "linear", couplage = (E=true, R=true))

Main.Genesys.Liion_electro_chimique — Type

Liion_electro_chimique

A mutable struct that represents a Li-ion battery aging model for State of Health (SoH) computation. This model is detailled in B. Xu et al. "Modeling of Lithium-Ion Battery Degradation for Cell Life Assessment," doi : 10.1109/TSG.2016.2578950, link.

The structure have a lot of parameters but most of them are set to default values.

Parameters:

α_p_ch::Float64: Charging maximum C-rate (default : 1.5)

α_p_dch::Float64: Discharging maximum C-rate (default : 1.5)

α_soc_min::Float64: Minimum threshold of charge (normalized) (default : 0.2)

α_soc_max::Float64: Maximum threshold of charge (normalized) (default : 0.8)

SoH_threshold::Float64: SoH level to replace the battery (default : 0.8)

couplage::NamedTuple: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling)

temperature::Float64: Temperature of the battery (no temprature model is implemented for the battery).

soc_model::String: Model name for State of Charge (SoC) computation. Available models are listed

calendar::Bool: Whether to include calendar aging in the SoH computation (default : true)

soc_ini::Float64: Initial State of Charge (SoC) for the beginning of the simulation (default : 0.5)

soh_ini::Float64: Initial State of Health (SoH) for the beginning of the simulation (default : 1)

update_by_year::Int64: Rainflow SoH computation by year (default : 12)

fatigue_data::DataFrames.DataFrame: DataFrame containing fatigue data (DoD, ncycle) (Default NMC battery data are provided)

Example of usage : Liion_electro_chimique(update_by_year = 12, soc_model = "linear", couplage = (E=true, R=true))

Main.Genesys.Liion_fixed_lifetime — Type

Liionfixedlifetime

A mutable struct that represents a Li-ion battery model with a linear calendar aging for State of Health (SoH) computation.

The structure have a lot of parameters but most of them are set to default values.

Parameters:

α_p_ch::Float64: Charging maximum C-rate (default : 1.5)

α_p_dch::Float64: Discharging maximum C-rate (default : 1.5)

α_soc_min::Float64: Minimum threshold of charge (normalized) (default : 0.2)

α_soc_max::Float64: Maximum threshold of charge (normalized) (default : 0.8)

SoH_threshold::Float64: SoH level to replace the battery (default : 0.8)

couplage::NamedTuple: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling)

soc_model::String: Model name for State of Charge (SoC) computation. Available models are listed

soc_ini::Float64: Initial State of Charge (SoC) for the beginning of the simulation (default : 0.5)

soh_ini::Float64: Initial State of Health (SoH) for the beginning of the simulation (default : 1)

Example of usage : Liion_fixed_lifetime(;soc_model = "polynomial", couplage = (E=true, R=true))

Main.Genesys.Liion_energy_exchanged — Type

Liion_energy_exchanged

A mutable struct that represents a Li-ion battery model Energy Throughput aging model for State of Health (SoH) computation.

The structure have a lot of parameters but most of them are set to default values.

Parameters:

α_p_ch::Float64: Charging maximum C-rate (default : 1.5)

α_p_dch::Float64: Discharging maximum C-rate (default : 1.5)

α_soc_min::Float64: Minimum threshold of charge (normalized) (default : 0.2)

α_soc_max::Float64: Maximum threshold of charge (normalized) (default : 0.8)

nCycle::Float64: Number of cycle before reaching EOL (should be found in the cycle to failure curve)

SoH_threshold::Float64: SoH level to replace the battery (default : 0.8)

couplage::NamedTuple: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling)

soc_model::String: Model name for State of Charge (SoC) computation. Available models are listed

calendar::Bool: Whether to include calendar aging in the SoH computation (default : true)

soc_ini::Float64: Initial State of Charge (SoC) for the beginning of the simulation (default : 0.5)

soh_ini::Float64: Initial State of Health (SoH) for the beginning of the simulation (default : 1)

Example of usage : Liion_energy_exchanged(;calendar = true, nCycle = fatigue_data.cycle[findfirst(fatigue_data.DoD .> (0.6))], soc_model = "polynomial", couplage = (E=true, R=true))

Here the nCycle is selected from the cycle to failure curve using 60% DoD.

Thermal Storage

Main.Genesys.ThermalStorage — Type

ThermalStorage

A mutable struct representing a thermal energy storage model with various parameters, initial conditions, variables, and an inner constructor.

Fields

α_p_ch::Float64: Maximum charging power factor

α_p_dch::Float64: Maximum discharging power factor

η_ch::Float64: Charging efficiency

η_dch::Float64: Discharging efficiency

η_self::Float64: Self-discharge rate

α_soc_min::Float64: Minimum state of charge factor

α_soc_max::Float64: Maximum state of charge factor

lifetime::Int64: Storage lifetime in years

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds of storage capacity

Erated_ini::Float64: Initial rated storage energy capacity

soc_ini::Float64: Initial state of charge

soh_ini::Float64: Initial state of health

Erated::AbstractArray{Float64,2}: Rated storage energy capacity over time and scenarios

carrier::Heat: Heat carrier for the thermal storage

soc::AbstractArray{Float64,3}: State of charge over time, years, and scenarios

cost::AbstractArray{Float64,2}: Cost of the thermal storage over time and scenarios

Hydrogen

Main.Genesys.H2Tank — Type

H2Tank

A mutable struct representing a hydrogen tank storage model with various parameters, initial conditions, variables, and an inner constructor.

Fields

α_p_ch::Float64: Maximum charging power factor (default: 1.5)

α_p_dch::Float64: Maximum discharging power factor (default: 1.5)

η_ch::Float64: Charging efficiency (default: 1.0)

η_dch::Float64: Discharging efficiency (default: 1.0)

η_self::Float64: Self-discharge rate (default: 0.0)

α_soc_min::Float64: Minimum state of charge factor (default: 0.0)

α_soc_max::Float64: Maximum state of charge factor (default: 1.0)

lifetime::Int64: Storage lifetime in years (default: 25)

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds of storage capacity (default: (lb = 0., ub = 10000.))

Erated_ini::Float64: Initial rated storage energy capacity (default: 1e-6)

soc_ini::Float64: Initial state of charge (default: 0.5)

soh_ini::Float64: Initial state of health (default: 1.0)

Erated::AbstractArray{Float64,2}: Rated storage energy capacity over time and scenarios

carrier::Hydrogen: Hydrogen carrier for the H2 tank

soc::AbstractArray{Float64,3}: State of charge over time, years, and scenarios

cost::AbstractArray{Float64,2}: Cost of the hydrogen storage over time and scenarios

Carrier

Main.Genesys.EnergyCarrier — Type

EnergyCarrier

An abstract type representing an energy carrier (e.g., Main.Genesys.Electricity, Main.Genesys.Heat, Main.Genesys.Hydrogen).

Main.Genesys.Electricity — Type

Electricity()

A mutable struct representing electricity as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with electricity.

Main.Genesys.Heat — Type

Heat()

A mutable struct representing heat as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with heat.

Main.Genesys.Hydrogen — Type

Hydrogen()

A mutable struct representing hydrogen as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with hydrogen.

Demand

Main.Genesys.Demand — Type

Demand(; carrier = Electricity())

A mutable struct representing energy demand in the grid. The struct is a subtype of AbstractDemand.

Fields

carrier::EnergyCarrier: The type of energy carrier for the demand (e.g., Electricity, Heat, etc.) this structre also contains the quantity for each timestamp. see Main.Genesys.EnergyCarrier

timestamp::Array{DateTime,3}: A 3-dimensional array of DateTime objects representing timestamps associated with the demand.

Example

```julia demand = Demand(carrier = Electricity())

outer