Structures and Functions

Microgrid

Genesys.Microgrid — Type

Microgrid()

Structure containing every element of the grid including immaterial ones like electrical demands.

Those elements are divided in 5 types and stored in 5 vectors with one field for each element of a type :

demands inheriting from AbstractDemand and containing a demand one or multiple Main.Genesys.EnergyCarrier (Electricity, Heat ,Hydrogen);

generations inheriting from AbstractGeneration and containing energy production/generation element;

storages inheriting from AbstractStorage and containing element responsible for the storage of different energy vectors;

converters inheriting from AbstractConverter and containing any element that transform an energy vector into another one;

grids inheriting from AbstractGrid and containing elements that sell or buy energy vectors.

(usually an external grid from which can be bought and sold energy);

These assets can be later added with the add!(mg::Microgrid, assets...) See also Main.Genesys.add! for a step by step declaration

Genesys.GlobalParameters — Type

mutable struct GlobalParameters

Structure containing global parameter for the simulation and optimization.

This struct is contained inside the Microgrid struct and provide global parameters for sizing and simulating.

Parameters :

(in sizing context) the number of operation and investment stages (ny, nh) or

(in simulation context) Time horizon and number of operation per year (ny, nh);

the number of scenario (ns);

optionnal : the yearly and hourly time steps (Δh, Δy) [default = (1,1)];

optionnal : the inlation rate (τ) [default = 0.045];

optionnal : the minimum renewable_share [default = 0.]

Storages

Batteries

Genesys.AbstractLiion — Type

abstract type AbstractLiion <: AbstractStorage end

An abstract type, parent of all Liion types

Genesys.Liion — Type

Liion

A mutable struct representing a Li-ion battery model with state of charge (SoC) computation and aging models.

Parameters

eff_model::AbstractLiionEffModel: Model for state of charge computation.

SoH_model::AbstractLiionAgingModel: Model for aging computation.

α_soc_min::Float64: Minimum threshold of charge (normalized).

α_soc_max::Float64: Maximum threshold of charge (normalized).

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds for the battery capacity.

SoH_threshold::Float64: State of health (SoH) level to replace the battery.

couplage::NamedTuple{(:E, :R), Tuple{Bool, Bool}}: Tuple to indicate if SoH should influence other parameters.

Initial Conditions

Erated_ini::Float64: Initial capacity of the battery in Wh.

soc_ini::Float64: Initial state of charge for the beginning of the simulation.

soh_ini::Float64: Initial state of health for the beginning of the simulation.

Variables

Erated::AbstractArray{Float64, 2}: Battery capacity.

carrier::Electricity: Type of energy.

soc::AbstractArray{Float64, 3}: 3-dimensional matrix (h, y, s) containing the state of charge [0-1].

soh::AbstractArray{Float64, 3}: 3-dimensional matrix (h, y, s) containing the state of health [0-1].

Eco

cost::AbstractArray{Float64, 2}: Economic cost.

Example

julia Liion(eff_model = PolynomialLiionEfficiency(), SoH_model = FixedLifetimeLiion())

Battery Efficiency Models

Genesys.AbstractLiionEffModel — Type

abstract type AbstractLiionEffModel end

An abstract type, parent of all Liion efficieny model types

Genesys.FixedLiionEfficiency — Type

FixedLiionEfficiency <: AbstractLiionEffModel

A mutable struct that represents a Li-ion battery efficiency model for SoC computation. This model implement a fixed efficiency that can decrease linearly with battery aging.

Parameters:

η_ch::Float64: Charging efficiency (default : 0.98)

η_dch::Float64: Discharging efficiency (default : 0.98)

α_p_ch::Float64: Maximum charging C-rate (default : 1.5)

α_p_dch::Float64: Maximum discharging C-rate (default : 1.5)

η_deg_coef::Float64: The efficiency degradation coefficient (default : 0.2303, ref : Redondo Iglesias - Efficiency Degradation Model of Lithium-Ion Batteries for Electric Vehicles)

couplage::NamedTuple{(:E, :R), Tuple{Bool, Bool}}: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling) (default : (E = true, R = true))

η_self::Float64: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling) (default : 0.0005)

Example

FixedLiionEfficiency()

Genesys.PolynomialLiionEfficiency — Type

PolynomialLiionEfficiency <: AbstractLiionEffModel

A mutable struct that represents a Li-ion battery efficiency model with polynomial efficiency characteristics for SoC computation. This model implements an efficiency model based on polynomial coefficients of the form ax² + bx + c that can be used to calculate efficiency.

Parameters:

a_η_ch::Float64: Coefficient 'a' for charging efficiency (default: 0.0033)

b_η_ch::Float64: Coefficient 'b' for charging efficiency (default: 0.0297)

c_η_ch::Float64: Coefficient 'c' for charging efficiency (default: 0.99814)

a_η_dch::Float64: Coefficient 'a' for discharging efficiency (default: 0.002232)

b_η_dch::Float64: Coefficient 'b' for discharging efficiency (default: 0.0246)

c_η_dch::Float64: Coefficient 'c' for discharging efficiency (default: 1)

η_deg_coef::Float64: The efficiency degradation coefficient (default: 0.2303, ref: Redondo Iglesias - Efficiency Degradation Model of Lithium-Ion Batteries for Electric Vehicles)

couplage::NamedTuple{(:E, :R), Tuple{Bool, Bool}}: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stands for capacity coupling and R for efficiency coupling) (default: (E = true, R = false))

α_p_ch::Float64: Maximum charging C-rate (default: 1.5)

α_p_dch::Float64: Maximum discharging C-rate (default: 1.5)

η_self::Float64: Auto-discharge factor (default: 0.0005)

Example

PolynomialLiionEfficiency()

Battery Aging Models

Genesys.AbstractLiionAgingModel — Type

abstract type AbstractLiionAgingModel end

An abstract type, parent of all Liion aging model types

Genesys.EnergyThroughputLiion — Type

EnergyThroughputLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries based on energy throughput, which accounts for calendar aging and cycling aging. This model calculates aging based on the cumulative energy throughput and additionnaly calendar aging.

Parameters:

calendar::Bool: A boolean value to indicate whether to consider calendar aging (default: true)

nCycle::Int64: An adjusted number of cycle to reach the SoH replacement threshold after nCycle_ini cycles (default: 6000)

nCycle_ini::Int64: The total number of cycles achievable before reaching EOL(default: 6000).

Δcal::Float64: The calendar aging parameter (default: 1 - exp(-4.14e-10 * 3600))

Example

julia EnergyThroughputLiion(;nCycle_ini = Int(floor(fatigue_data.cycle[findfirst(fatigue_data.DoD .> (0.6))])))

This set the feasible number of cycles corresponding to the number of 60% DoD cycles in the cycle-to-failure curve (named fatigue_data). Supposing that fatigue data is a dataframe with a cycle column and a DoD column

Genesys.FixedLifetimeLiion — Type

FixedLifetimeLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries with a fixed lifetime in years.

Parameters:

lifetime::Int64: The fixed lifetime of the battery in years (default: 12)

Example

FixedLifetimeLiion()

Genesys.RainflowLiion — Type

RainflowLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries based on rainflow cycle counting, considering calendar aging and cycling aging.

Parameters:

calendar::Bool: A boolean value for the activation of calendar aging (default: true)

update_by_year::Int64: Number of updates each year (default: 12)

fatigue_data::DataFrames.DataFrame: Cycle to failure curve data (default: loaded from a CSV file with a DoD column and a nCycle column)

Example

julia RainflowLiion(fatigue_data = fatigue_data)

supposing that fatigue_data is a dataframe with a cycle column and a DoD column

Genesys.SemiEmpiricalLiion — Type

SemiEmpiricalLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries based on a semi-empirical model, accounting for temperature and other parameters.

Parameters:

update_by_year::Int64: Number of updates each year (default: 12)

Parameters for NMC bateries (as in Technoeconomic model of second-life batteries for utility-scale solar considering calendar and cycle aging Ian Mathews, Bolun Xu, Wei He, Vanessa Barreto, Tonio Buonassisi, Ian Marius Peters):

alpha_sei::Float64

beta_sei::Float64

k_delta1::Float64

k_delta2::Float64

k_sigma::Float64

sigma_ref::Float64

k_T::Float64

T_ref::Float64

k_t::Float64

temperature::Float64: Temperature of the battery in Kelvin (default: 298)

Sum_fd::AbstractArray{Float64,1}: Used as a memory of the cumulated fatigue of the battery

Example

SemiEmpiricalLiion()

Thermal Storage

Genesys.ThermalStorage — Type

mutable struct ThermalStorage <: AbstractStorage

A mutable struct representing a thermal energy storage model with various parameters, initial conditions, variables, and an inner constructor.

Parameters

α_p_ch::Float64: Maximum charging power factor

α_p_dch::Float64: Maximum discharging power factor

eff_model::AbstractTESEffModel: Efficiency model

α_soc_min::Float64: Minimum state of charge factor

α_soc_max::Float64: Maximum state of charge factor

SoH_model::AbstractTESAgingModel: Storage aging model (fixed lifetime in years)

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds of storage capacity

Erated_ini::Float64: Initial rated storage energy capacity

soc_ini::Float64: Initial state of charge

soh_ini::Float64: Initial state of health

Erated::AbstractArray{Float64,2}: Rated storage energy capacity over time and scenarios

carrier::Heat: Heat carrier for the thermal storage

soc::AbstractArray{Float64,3}: State of charge over time, years, and scenarios

cost::AbstractArray{Float64,2}: Cost of the thermal storage over time and scenarios

Hydrogen

Genesys.H2Tank — Type

mutable struct H2Tank <: AbstractStorage

A mutable struct representing a hydrogen tank storage model with various parameters, initial conditions, variables, and an inner constructor.

Parameters

α_p_ch::Float64: Maximum charging power factor (default: 1.5)

α_p_dch::Float64: Maximum discharging power factor (default: 1.5)

η_ch::Float64: Charging efficiency (default: 1.0)

η_dch::Float64: Discharging efficiency (default: 1.0)

η_self::Float64: Self-discharge rate (default: 0.0)

α_soc_min::Float64: Minimum state of charge factor (default: 0.0)

α_soc_max::Float64: Maximum state of charge factor (default: 1.0)

lifetime::Int64: Storage lifetime in years (default: 25)

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds of storage capacity (default: (lb = 0., ub = 10000.))

Erated_ini::Float64: Initial rated storage energy capacity (default: 1e-6)

soc_ini::Float64: Initial state of charge (default: 0.5)

soh_ini::Float64: Initial state of health (default: 1.0)

Erated::AbstractArray{Float64,2}: Rated storage energy capacity over time and scenarios

carrier::Hydrogen: Hydrogen carrier for the H2 tank

soc::AbstractArray{Float64,3}: State of charge over time, years, and scenarios

cost::AbstractArray{Float64,2}: Cost of the hydrogen storage over time and scenarios

Generation

Genesys.Solar — Type

mutable struct Solar <: AbstractGeneration

A mutable struct representing a solar power generation source. It contains information about the lifetime, bounds, initial conditions, variables, and costs related to the solar generation.

Parameters

-SoH_model::AbstractPVAgingModel (default: FixedLifetimeSolar(25)): The aging model with a fixed lifetime of the solar power generation system. -bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}} (default: (lb = 0., ub = 1000.)): The lower and upper bounds of the solar power generation system. -powerMax_ini::Float64 (default: 0.): The initial maximum power output of the solar power generation system. -soh_ini::Float64 (default: 1.): The initial state of health of the solar power generation system.

Carrier

Genesys.EnergyCarrier — Type

abstract type EnergyCarrier end

An abstract type, parent of all energy carrier types (electricity, heat, hydrogen...)

Genesys.Electricity — Type

Electricity

A mutable struct representing electricity as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with electricity.

Example

Electricity()

Genesys.Heat — Type

Heat

A mutable struct representing heat as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with heat.

Example

Heat()

Genesys.Hydrogen — Type

Hydrogen

A mutable struct representing hydrogen as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with hydrogen.

Hydrogen()

Converter

Genesys.Electrolyzer — Type

Genesys.FuelCell — Type

FuelCell represents the main structure for modeling fuel cell behavior, including efficiency, aging, and other key parameters.

Fields

eff_model::AbstractFuelCellEffModel: The efficiency model used to calculate fuel cell efficiency.

SoH_model::AbstractFuelCellAgingModel: The aging model used to track the State of Health (SoH) of the fuel cell over time.

couplage::Bool: Indicates whether the SoH should influence other parameters (e.g., efficiency, output voltage).

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Bounds for operational parameters, such as voltage or power.

SoH_threshold::Float64: SoH level at which the fuel cell should be replaced.

V_J_ini::AbstractArray{Float64,2}: Initial voltage-current relationship for the fuel cell.

soh_ini::Float64: Initial SoH at the start of the simulation (default: 1.0, corresponding to 100% health).

N_cell::Int64: Number of cells assembled in the fuel cell stack.

surface::Float64: Surface area of the fuel cell stack (unit should be documented if relevant).

η::AbstractArray{Float64,3}: Efficiency evolution over time (hydrogen to electricity).

carrier::Vector{EnergyCarrier}: Vector of energy carriers used by the fuel cell (e.g., hydrogen, oxygen).

soh::AbstractArray{Float64,3}: Time-evolving State of Health for the fuel cell.

cost::AbstractArray{Float64,2}: Economic metrics associated with fuel cell operation (e.g., cost per hour, maintenance).

Constructor Parameters

eff_model::AbstractFuelCellEffModel=PolarizationFuelCellEfficiency(): Efficiency model to be used.

SoH_model::AbstractFuelCellAgingModel=PowerAgingFuelCell(): Aging model to be used.

couplage::Bool=true: Whether SoH influences other parameters.

bounds=(lb=0.0, ub=50.0): sizing power bounds for all-in-one design.

SoH_threshold::Float64=0.9: SoH replacement threshold (default: 90%).

V_J_ini::AbstractArray{Float64,2}: Initial voltage-current data (default loaded from a CSV file).

soh_ini::Float64=1.0: Initial SoH at the start of the simulation.

Example

fuel_cell = FuelCell(eff_model=PolarizationFuelCellEfficiency(),
                     SoH_model=PowerAgingFuelCell(),
                     couplage=true,
                     bounds=(lb=0.0, ub=60.0),
                     SoH_threshold=0.8,
                     soh_ini=1.0)

This initializes a FuelCell structure with custom sizing power bounds and a lower SoH replacement threshold.

Genesys.Heater — Type

mutable struct Heater <: AbstractConverter

A mutable struct representing a Heater, which is a subtype of AbstractConverter.

Parameters

eff_model::AbstractHeaterEffModel: The model for conversion efficiency of the heater.

SoH_model::AbstractHeaterAgingModel: The aging model for the heater. The default model use a fixed expected lifetime for the heater (in years).

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: The lower and upper bounds of the heater's power.

powerMax_ini::Float64: The initial maximum power output of the heater.

soh_ini::Float64: The initial state of health of the heater.

Variables

powerMax::AbstractArray{Float64,2}: The maximum power output of the heater.

carrier::Vector{Any}: The carrier vector.

cost::AbstractArray{Float64,2}: The cost associated with the heater.

Example

heater = Heater(eff_model=ConstEfficiencyHeater(η_E_H=1.0), SoH_model=FixedLifetimeHeater(lifetime=25), bounds=(lb=30.0, ub=30.0), powerMax_ini=30.0, soh_ini=1.0)

Demand

Genesys.Demand — Type

Demand(; carrier = Electricity())

A mutable struct representing energy demand in the grid. The struct is a subtype of AbstractDemand.

Parameters

carrier::EnergyCarrier: The type of energy carrier for the demand (e.g., Electricity, Heat, etc.) this structre also contains the quantity for each timestamp. see Main.Genesys.EnergyCarrier

timestamp::Array{DateTime,3}: A 3-dimensional array of DateTime objects representing timestamps associated with the demand.

Example

demand = Demand(carrier = Electricity())

grid

Genesys.Grid — Type

mutable struct Grid <: AbstractGrid

A mutable struct representing a Grid, which is a subtype of AbstractGrid.

Parameters

powerMax::Float64: The maximum power capacity of the grid.

Variables

carrier::EnergyCarrier: The energy carrier associated with the grid, such as electricity.

cost_in::AbstractArray{Float64,3}: The input cost associated with the grid.

cost_out::AbstractArray{Float64,3}: The output cost associated with the grid.

cost_exceed::AbstractArray{Float64,2}: The cost associated with exceeding the grid's capacity.

Example

grid = Grid(powerMax=36.0, carrier=Electricity())

Scenario

Genesys.Scenarios — Type

function Scenarios(mg::Microgrid, d::Dict{})

Constructor function for creating a new Scenarios instance based on a given Microgrid and a Dict containing scenario data.

Arguments

mg::Microgrid: A Microgrid instance.

d::Dict{}: A dictionary containing scenario data.

Returns

Scenarios: A Scenarios instance with the specified data.

Example

microgrid = ...
scenario_data = ...

scenarios = Scenarios(microgrid, scenario_data)

function Scenarios(mg::Microgrid, d::Dict{}, same_year::Bool; seed = [])

Constructor function for creating a new Scenarios instance based on a given Microgrid and a Dict containing scenario data. It allows for repetitive years for longer scenarios.

Arguments

mg::Microgrid: A Microgrid instance.

d::Dict{}: A dictionary containing scenario data.

same_year::Bool : If true, the function repeats the same year for all years in the microgrid.

seed::Array=[]: An array specifying the seed for selecting the scenario number and offer reproductivity.

Returns

Scenarios: A Scenarios instance with the specified data.

Example

microgrid = ...
scenario_data = ...

scenarios = Scenarios(microgrid, scenario_data, true; seed = [1, 2, 3])

Scenarios Reducer

Genesys.ManualReducer — Type

mutable struct ManualReducer <: AbstractScenariosReducer
    h::Union{UnitRange{Int64}, Int64}
    y::Union{UnitRange{Int64}, Int64}
    s::Union{UnitRange{Int64}, Int64}
end

A mutable structure representing a manual scenario reduction. Inherits from the AbstractScenariosReducer abstract type.

Fields

h::Union{UnitRange{Int64}, Int64}: The hours to be included in the reduced scenario.

y::Union{UnitRange{Int64}, Int64}: The years to be included in the reduced scenario.

s::Union{UnitRange{Int64}, Int64}: The scenarios to be included in the reduced scenario.

Constructor

ManualReducer(; h = 1:8760, y = 2:3, s = 1:1) = new(h, y, s)

Genesys.SAAReducer — Type

mutable struct SAAReducer <: AbstractScenariosReducer
    nsample::Int64
end

A mutable structure representing a Sample Average Approximation (SAA) scenario reduction. Inherits from the AbstractScenariosReducer abstract type.

Fields

nsample::Int64: The number of samples to be included in the reduced scenario.

Constructor

SAAReducer(; nsample = 100) = new(nsample)

Genesys.MeanValueReducer — Type

mutable struct MeanValueReducer <: AbstractScenariosReducer
end

A mutable structure representing an Expected Value (Mean Value) scenario reduction. Inherits from the AbstractScenariosReducer abstract type.

Constructor

MeanValueReducer() = new()

Genesys.FeatureBasedReducer — Type

mutable struct FeatureBasedReducer <: AbstractScenariosReducer
    transformation::Union{UnionAll, Nothing}
    reduction::Union{AbstractDimensionReducer, Nothing}
    clustering::AbstractClusteringMethod
end

A mutable structure representing a Clustering scenario reduction using feature-based methods. Inherits from the AbstractScenariosReducer abstract type.

Fields

transformation: A transformation method to apply before clustering (default: UnitRangeTransform)

reduction: A dimensionality reduction method to apply before clustering (default: StatsReduction())

clustering: The clustering method to use (default: KmedoidsClustering())

Constructor

FeatureBasedReducer(; transformation = UnitRangeTransform,
                    reduction = StatsReduction(),
                    clustering = KmedoidsClustering()) = new(transformation, reduction, clustering)

Dimention Reducer

Genesys.PCAReduction — Type

mutable struct PCAReduction <: AbstractDimensionReducer
    n_components::Int64
end

A mutable structure representing Principal Component Analysis (PCA) dimensionality reduction. Inherits from the AbstractDimensionReducer abstract type.

Fields

n_components: The number of principal components to use for dimensionality reduction (default: 2)

Constructor

PCAReduction(; n_components = 2) = new(n_components)

Genesys.StatsReduction — Type

struct StatsReduction <: AbstractDimensionReducer

A structure representing statistics-based dimensionality reduction. Inherits from the AbstractDimensionReducer abstract type.

Constructor

StatsReduction()

Transformation

StatsBase.UnitRangeTransform — Type

Unit range normalization

StatsBase.ZScoreTransform — Type

Standardization (Z-score transformation)

Custering

Genesys.KmedoidsClustering — Type

mutable struct KmedoidsClustering <: AbstractClusteringMethod

A structure representing K-medoids clustering method. Inherits from the AbstractClusteringMethod abstract type.

Fields

n_clusters::Int64: Number of clusters to form.

distance::Distances.SemiMetric: Distance metric to use for calculating the dissimilarity between points.

log::Bool: Whether to print logs during the clustering process.

Constructor

KmedoidsClustering(; n_clusters = 20, distance = Distances.Euclidean(), log = true)

Generator

Genesys.MarkovGenerator — Type

mutable struct MarkovGenerator <: AbstractScenariosGenerator

A structure representing a Markov chain-based scenario generator.

Fields

nstate::Int64: Number of states, which corresponds to the number of clusters per hour per month.

algo::String: Algorithm used for clustering the states.

markovchains::NamedTuple{(:wk, :wkd), Tuple{MarkovChain, MarkovChain}}: A named tuple containing the Markov chains for weekdays and weekends.

Example

markov_generator = MarkovGenerator(nstate=10, algo="kmeans")

Genesys.AnticipativeGenerator — Type

mutable struct AnticipativeGenerator <: AbstractScenariosGenerator

A structure representing an anticipative scenario generator (perfect foresight).

Fields

forecast: The field to store perfect foresight information.

Example

anticipative_generator = AnticipativeGenerator()

Base generic function

Generate

Genesys.generate — Function

Generate scenarios using a Markov chain.

Parameters: generator: a MarkovGenerator object s0: a ny x generator.nstate matrix representing the initial state t0: a DateTime object representing the initial timestamp nstep: an integer representing the number of steps to generate ny: an integer representing the number of scenarios to generate ns: an integer representing the number of seeds for the generator

Returns:

a list of ny ny x ns x nstep matrix representing the generated scenarios

a ny x ns matrix representing the probability associated with each generated scenario

generate(generator::AnticipativeGenerator, s0, t0::DateTime, nstep::Int64; ny::Int64=1, ns::Int64=1, h::Int64=1)

Generate perfect forecast.

Arguments

generator::AnticipativeGenerator: Anticipative generator struct.

s0: Vector of initial state(s).

t0::DateTime: Initial time.

nstep::Int64: Number of steps to simulate.

ny::Int64=1: Number of outputs.

ns::Int64=1: Number of scenarios.

h::Int64=1: Initial time step.

Returns

forecast: Matrix of shape (nstep, ny, ns) with the forecasted values.

probs: Array of shape (ny, ns) with the probability distribution of each scenario.

Reduce

Genesys.reduce — Function

function reduce(reducer::ManualReducer, ω::Scenarios)

Reduces the given Scenarios object using the specified ManualReducer.

Arguments

reducer::ManualReducer: The reducer used to specify the reduction process.

ω::Scenarios: The Scenarios object to be reduced.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario.

Example

manual_reducer = ManualReducer(h = 1:8760, y = 2:3, s = 1:1)
reduced_scenarios, probabilities = reduce(manual_reducer, scenarios)

function reduce(reducer::SAAReducer, ω::Scenarios; y::Int64 = 1, s::Int64 = 1)

Reduces the given Scenarios object using the specified SAAReducer, implementing the Sample Average Approximation scenario reduction method.

Arguments

reducer::SAAReducer: The reducer used to specify the reduction process.

ω::Scenarios: The Scenarios object to be reduced.

Keyword Arguments

y::Int64 = 1: The initial year to be considered.

s::Int64 = 1: The initial scenario to be considered.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario.

Example

saa_reducer = SAAReducer(nsample = 100)
reduced_scenarios, probabilities = reduce(saa_reducer, scenarios)

function reduce(reducer::MeanValueReducer, ω::Scenarios; y::Int64 = 1, s::Int64 = 1)

Reduces the given Scenarios object using the specified MeanValueReducer, implementing the Expected Value scenario reduction method.

Arguments

reducer::MeanValueReducer: The reducer used to specify the reduction process.

ω::Scenarios: The Scenarios object to be reduced.

Keyword Arguments

y::Int64 = 1: The initial year to be considered.

s::Int64 = 1: The initial scenario to be considered.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario (in this case, a single element array containing 1).

Example

mean_value_reducer = MeanValueReducer()
reduced_scenarios, probabilities = reduce(mean_value_reducer, scenarios)

function reduce(reducer::FeatureBasedReducer, ω::Scenarios; y::Int64 = 1, s::Int64 = 1)

Reduces the given Scenarios object using the specified FeatureBasedReducer, implementing a clustering-based scenario reduction method.

Arguments

reducer::FeatureBasedReducer: The reducer used to specify the reduction process, including transformation, dimension reduction, and clustering methods.

ω::Scenarios: The Scenarios object to be reduced.

Keyword Arguments

y::Int64 = 1: The initial year to be considered.

s::Int64 = 1: The initial scenario to be considered.

Returns

A new reduced Scenarios object.

A probability vector representing the likelihood of each scenario in the reduced set.

An array of assignments indicating which cluster each original scenario belongs to.

Example

feature_based_reducer = FeatureBasedReducer(transformation = UnitRangeTransform,
                                            reduction = StatsReduction(),
                                            clustering = KmedoidsClustering())
reduced_scenarios, probabilities, assignments = reduce(feature_based_reducer, scenarios)