Introduction to Genesys

Welcome to the Genesys module, a powerful tool for simulating and optimizing microgrids.

In this page, we'll provide an overview of the main components and features of the Genesys module.

Main Components

The Genesys module includes the following main components:

Assets: Microgrid components, such as demand, solar, liion, thermal storage, and grid.

Scenarios: Tools for generating and reducing scenarios for simulations.

Optimization: Design and operation optimization tools, including controllers and designers.

Simulation: Functions for simulating microgrids over different time horizons and scenarios.

Utils: Utilities for metrics, plotting, and data saving.

Key Features

The Genesys module offers a wide range of features for microgrid simulations and optimizations:

A flexible and extensible structure that allows for the incorporation of different asset types and optimization algorithms.

A comprehensive set of optimization tools, including dummy, anticipative, rule-based control, and online learning for frequency control.

Scenario generation and reduction techniques for efficient and accurate simulations.

Built-in plotting and metrics utilities for easy analysis and visualization of results.

To learn more about how to use the Genesys module, please refer to the API Reference and other pages in this documentation.

Functions

Main.Genesys.GlobalParameters — Type

GlobalParameters(nh, ny, ns)

Structure containing global parameter for the simulation and optimization.

This struct is contained inside the Microgrid struct and provide global parameters for sizing and simulating. Parameters are :

(in sizing context) the number of operation and investment stages (ny, nh) or

(in simulation context) Time horizon and number of operation per year (ny, nh);

the number of scenario (ns);

optionnal : the yearly and hourly time steps (Δh, Δy) [default = (1,1)];

optionnal : the inlation rate (τ) [default = 0.045];

optionnal : the minimum renewable_share [default = 0.]

Main.Genesys.Microgrid — Type

Microgrid()

Structure containing every element of the grid including immaterial ones like electrical demands.

Those elements are divided in 5 types and stored in 5 vectors with one field for each element of a type :

demands inheriting from AbstractDemand and containing a demand for each energy (Electricity and Heat);

generations inheriting from AbstractGeneration and containing energy

production/generation assets;

storages inheriting from AbstractStorage and containing energy storage assets;

converters inheriting from AbstractConverter and containing energy convertion assets;

grids inheriting from AbstractGrid and containing energy market assets

(usually an external grid from which can be bought and sold energy);

These assets can be later added with the add!(mg::Microgrid, assets...) See also Main.Genesys.add! for a step by step declaration

Main.Genesys.Options — Type

Options

A mutable struct representing the options for the simulation.

Main.Genesys.Solar — Type

Solar <: AbstractGeneration

A mutable struct representing a solar power generation source. It contains information about the lifetime, bounds, initial conditions, variables, and costs related to the solar generation.

Parameters -lifetime::Int64 (default: 25): The lifetime of the solar power generation system. -bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}} (default: (lb = 0., ub = 1000.)): The lower and upper bounds of the solar power generation system. -powerMax_ini::Float64 (default: 0.): The initial maximum power output of the solar power generation system. -soh_ini::Float64 (default: 1.): The initial state of health of the solar power generation system.

Main.Genesys.add! — Method

add!(mg::Microgrid, assets...)

Add to the given microgrid mg every asset listed as parameter into the appropriate vector.

Meanwhile data struct are created, they are sized according to the number of scenario, years, and hours per year, ns, ny, nh resulting in 3D arrays

Main.Genesys.compute_investment_dynamics! — Method

compute_investment_dynamics!(y::Int64, s::Int64, h2tank::H2Tank, decision::Union{Float64, Int64})

Compute the investment dynamics of the hydrogen tank storage for a given year (y) and scenario (s), using the provided decision.

Main.Genesys.compute_investment_dynamics! — Method

compute_investment_dynamics!(y::Int64, s::Int64, pv::Solar, decision::Union{Float64, Int64})

Compute the investment dynamics of the solar power generation system for a given year (y) and scenario (s), using the provided decision.

Main.Genesys.compute_investment_dynamics! — Method

compute_investment_dynamics!(y::Int64, s::Int64, tes::ThermalStorage, decision::Union{Float64, Int64})

Compute the investment dynamics of the thermal energy storage for a given year (y) and scenario (s), using the given decision variable.

Main.Genesys.compute_investment_dynamics — Method

computeinvestmentdynamics(h2tank::H2Tank, state::NamedTuple{(:Erated, :soc), Tuple{Float64, Float64}}, decision::Union{Float64, Int64})

Compute the investment dynamics of the hydrogen tank storage for the given state and decision.

Main.Genesys.compute_investment_dynamics — Method

computeinvestmentdynamics(pv::Solar, state::NamedTuple{(:powerMax,), Tuple{Float64}}, decision::Union{Float64, Int64})

Compute the investment dynamics of the solar power generation system for the given state and decision.

Main.Genesys.compute_investment_dynamics — Method

compute_operation_dynamics(tes::ThermalStorage, state::NamedTuple{(:Erated, :soc), Tuple{Float64, Float64}}, decision::Float64, Δh::Int64)

Compute the operation dynamics of the thermal energy storage, given the current state (Erated and soc), decision variable, and time step (Δh).

Main.Genesys.compute_operation_dynamics! — Method

compute_operation_dynamics!(h::Int64, y::Int64, s::Int64, h2tank::H2Tank, decision::Float64, Δh::Int64)

Compute the operation dynamics of the hydrogen tank storage for a given hour (h), year (y), and scenario (s), using the provided decision and time step (Δh).

Main.Genesys.compute_operation_dynamics! — Method

compute_operation_dynamics!(h::Int64, y::Int64, s::Int64, tes::ThermalStorage, decision::Float64, Δh::Int64)

Compute the operation dynamics of the thermal energy storage for a given hour (h), year (y), and scenario (s), using the given decision variable and time step (Δh).

Main.Genesys.compute_operation_dynamics — Method

compute_operation_dynamics(h2tank::H2Tank, state::NamedTuple{(:Erated, :soc), Tuple{Float64, Float64}}, decision::Float64, Δh::Int64)

Compute the operation dynamics of the hydrogen tank storage for the given state, decision, and time step (Δh).

Main.Genesys.isin — Method

isin(field::Vector, type::DataType)

Find if the datatype is in a mg field and where it is

For a given vector field and a Datatype type, if an object of the type type is in field return true and the index of the type in the vector else return false and NaN.

Main.Genesys.preallocate! — Method

preallocate!(h2tank::H2Tank, nh::Int64, ny::Int64, ns::Int64)

Preallocate memory for the H2Tank object with given dimensions for hours (nh), years (ny), and scenarios (ns).

Main.Genesys.preallocate! — Method

preallocate!(mg::Microgrid, controller::AbstractController)

Preallocate abstract controller

Data struct are sized according to the number of scenario, years and hours per year to store for each element to size a decision per operation stage.

Main.Genesys.preallocate! — Method

preallocate!(mg::Microgrid, designer::AbstractDesigner)

Preallocate abstract designer

Data struct are sized according to the number of scenario and years to store for each element to size a decision per investment stage.

Main.Genesys.preallocate! — Method

preallocate!(pv::Solar, nh::Int64, ny::Int64, ns::Int64)

Preallocate necessary arrays for the solar power generation model with the specified number of hours (nh), years (ny), and scenarios (ns).

Main.Genesys.preallocate! — Method

preallocate!(tes::ThermalStorage, nh::Int64, ny::Int64, ns::Int64)

Preallocate memory for the ThermalStorage object with given dimensions for hours (nh), years (ny), and scenarios (ns).

Main.Genesys.simulate! — Method

simulate!(h::Int64, y::Int64, s::Int64, mg::Microgrid, controller::AbstractController, designer::AbstractDesigner, ω_simu::Scenarios, options::Options)

Simulates the microgrid for a single hour within a given year and scenario.

Main.Genesys.simulate! — Method

simulate!(y::Int64, s::Int64, mg::Microgrid, controller::AbstractController, designer::AbstractDesigner, ω_simu::Scenarios, options::Options)

Simulates the microgrid for a single year within a given scenario.

Main.Genesys.simulate! — Method

simulate!(s::Int64, mg::Microgrid, controller::AbstractController, designer::AbstractDesigner, ω_simu::Scenarios, options::Options)

Simulates the microgrid for a single scenario.

Main.Genesys.simulate! — Method

simulate!(mg::Microgrid, controller::AbstractController, designer::AbstractDesigner, ω_simu::Scenarios; options::Options = Options())

Main simulation function. Simulates the microgrid over multiple scenarios based on the given options.