Assets

Storages

Batteries

Genesys.AbstractLiion — Type

abstract type AbstractLiion <: AbstractStorage end

An abstract type, parent of all Liion types

Genesys.Liion — Type

Liion

A mutable struct representing a Li-ion battery model with state of charge (SoC) computation and aging models.

Parameters

eff_model::AbstractLiionEffModel: Model for state of charge computation.

SoH_model::AbstractLiionAgingModel: Model for aging computation.

α_soc_min::Float64: Minimum threshold of charge (normalized).

α_soc_max::Float64: Maximum threshold of charge (normalized).

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds for the battery capacity.

SoH_threshold::Float64: State of health (SoH) level to replace the battery.

couplage::NamedTuple{(:E, :R), Tuple{Bool, Bool}}: Tuple to indicate if SoH should influence other parameters.

Initial Conditions

Erated_ini::Float64: Initial capacity of the battery in Wh.

soc_ini::Float64: Initial state of charge for the beginning of the simulation.

soh_ini::Float64: Initial state of health for the beginning of the simulation.

Variables

Erated::AbstractArray{Float64, 2}: Battery capacity.

carrier::Electricity: Type of energy.

soc::AbstractArray{Float64, 3}: 3-dimensional matrix (h, y, s) containing the state of charge [0-1].

soh::AbstractArray{Float64, 3}: 3-dimensional matrix (h, y, s) containing the state of health [0-1].

Eco

cost::AbstractArray{Float64, 2}: Economic cost.

Example

julia Liion(eff_model = PolynomialLiionEfficiency(), SoH_model = FixedLifetimeLiion())

Battery Efficiency Models

Genesys.AbstractLiionEffModel — Type

abstract type AbstractLiionEffModel end

An abstract type, parent of all Liion efficieny model types

Genesys.FixedLiionEfficiency — Type

FixedLiionEfficiency <: AbstractLiionEffModel

A mutable struct that represents a Li-ion battery efficiency model for SoC computation. This model implement a fixed efficiency that can decrease linearly with battery aging.

Parameters:

η_ch::Float64: Charging efficiency (default : 0.98)

η_dch::Float64: Discharging efficiency (default : 0.98)

α_p_ch::Float64: Maximum charging C-rate (default : 1.5)

α_p_dch::Float64: Maximum discharging C-rate (default : 1.5)

η_deg_coef::Float64: The efficiency degradation coefficient (default : 0.2303, ref : Redondo Iglesias - Efficiency Degradation Model of Lithium-Ion Batteries for Electric Vehicles)

couplage::NamedTuple{(:E, :R), Tuple{Bool, Bool}}: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling) (default : (E = true, R = true))

η_self::Float64: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stand for capacity coupling and R for efficiency coupling) (default : 0.0005)

Example

FixedLiionEfficiency()

Genesys.PolynomialLiionEfficiency — Type

PolynomialLiionEfficiency <: AbstractLiionEffModel

A mutable struct that represents a Li-ion battery efficiency model with polynomial efficiency characteristics for SoC computation. This model implements an efficiency model based on polynomial coefficients of the form ax² + bx + c that can be used to calculate efficiency.

Parameters:

a_η_ch::Float64: Coefficient 'a' for charging efficiency (default: 0.0033)

b_η_ch::Float64: Coefficient 'b' for charging efficiency (default: 0.0297)

c_η_ch::Float64: Coefficient 'c' for charging efficiency (default: 0.99814)

a_η_dch::Float64: Coefficient 'a' for discharging efficiency (default: 0.002232)

b_η_dch::Float64: Coefficient 'b' for discharging efficiency (default: 0.0246)

c_η_dch::Float64: Coefficient 'c' for discharging efficiency (default: 1)

η_deg_coef::Float64: The efficiency degradation coefficient (default: 0.2303, ref: Redondo Iglesias - Efficiency Degradation Model of Lithium-Ion Batteries for Electric Vehicles)

couplage::NamedTuple{(:E, :R), Tuple{Bool, Bool}}: Named tuple with two boolean values to indicate if the SoH should influence the other parameters (E stands for capacity coupling and R for efficiency coupling) (default: (E = true, R = false))

α_p_ch::Float64: Maximum charging C-rate (default: 1.5)

α_p_dch::Float64: Maximum discharging C-rate (default: 1.5)

η_self::Float64: Auto-discharge factor (default: 0.0005)

Example

PolynomialLiionEfficiency()

Battery Aging Models

Genesys.AbstractLiionAgingModel — Type

abstract type AbstractLiionAgingModel end

An abstract type, parent of all Liion aging model types

Genesys.EnergyThroughputLiion — Type

EnergyThroughputLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries based on energy throughput, which accounts for calendar aging and cycling aging. This model calculates aging based on the cumulative energy throughput and additionnaly calendar aging.

Parameters:

calendar::Bool: A boolean value to indicate whether to consider calendar aging (default: true)

nCycle::Int64: An adjusted number of cycle to reach the SoH replacement threshold after nCycle_ini cycles (default: 6000)

nCycle_ini::Int64: The total number of cycles achievable before reaching EOL(default: 6000).

Δcal::Float64: The calendar aging parameter (default: 1 - exp(-4.14e-10 * 3600))

Example

julia EnergyThroughputLiion(;nCycle_ini = Int(floor(fatigue_data.cycle[findfirst(fatigue_data.DoD .> (0.6))])))

This set the feasible number of cycles corresponding to the number of 60% DoD cycles in the cycle-to-failure curve (named fatigue_data). Supposing that fatigue data is a dataframe with a cycle column and a DoD column

Genesys.FixedLifetimeLiion — Type

FixedLifetimeLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries with a fixed lifetime in years.

Parameters:

lifetime::Int64: The fixed lifetime of the battery in years (default: 12)

Example

FixedLifetimeLiion()

Genesys.RainflowLiion — Type

RainflowLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries based on rainflow cycle counting, considering calendar aging and cycling aging.

Parameters:

calendar::Bool: A boolean value for the activation of calendar aging (default: true)

update_by_year::Int64: Number of updates each year (default: 12)

fatigue_data::DataFrames.DataFrame: Cycle to failure curve data (default: loaded from a CSV file with a DoD column and a nCycle column)

Example

julia RainflowLiion(fatigue_data = fatigue_data)

supposing that fatigue_data is a dataframe with a cycle column and a DoD column

Genesys.SemiEmpiricalLiion — Type

SemiEmpiricalLiion <: AbstractLiionAgingModel

A mutable struct that represents an aging model for Li-ion batteries based on a semi-empirical model, accounting for temperature and other parameters.

Parameters:

update_by_year::Int64: Number of updates each year (default: 12)

Parameters for NMC bateries (as in Technoeconomic model of second-life batteries for utility-scale solar considering calendar and cycle aging Ian Mathews, Bolun Xu, Wei He, Vanessa Barreto, Tonio Buonassisi, Ian Marius Peters):

alpha_sei::Float64

beta_sei::Float64

k_delta1::Float64

k_delta2::Float64

k_sigma::Float64

sigma_ref::Float64

k_T::Float64

T_ref::Float64

k_t::Float64

temperature::Float64: Temperature of the battery in Kelvin (default: 298)

Sum_fd::AbstractArray{Float64,1}: Used as a memory of the cumulated fatigue of the battery

Example

SemiEmpiricalLiion()

Thermal Storage

Genesys.ThermalStorage — Type

mutable struct ThermalStorage <: AbstractStorage

A mutable struct representing a thermal energy storage model with various parameters, initial conditions, variables, and an inner constructor.

Parameters

α_p_ch::Float64: Maximum charging power factor

α_p_dch::Float64: Maximum discharging power factor

eff_model::AbstractTESEffModel: Efficiency model

α_soc_min::Float64: Minimum state of charge factor

α_soc_max::Float64: Maximum state of charge factor

SoH_model::AbstractTESAgingModel: Storage aging model (fixed lifetime in years)

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds of storage capacity

Erated_ini::Float64: Initial rated storage energy capacity

soc_ini::Float64: Initial state of charge

soh_ini::Float64: Initial state of health

Erated::AbstractArray{Float64,2}: Rated storage energy capacity over time and scenarios

carrier::Heat: Heat carrier for the thermal storage

soc::AbstractArray{Float64,3}: State of charge over time, years, and scenarios

cost::AbstractArray{Float64,2}: Cost of the thermal storage over time and scenarios

Hydrogen

Genesys.H2Tank — Type

mutable struct H2Tank <: AbstractStorage

A mutable struct representing a hydrogen tank storage model with various parameters, initial conditions, variables, and an inner constructor.

Parameters

α_p_ch::Float64: Maximum charging power factor (default: 1.5)

α_p_dch::Float64: Maximum discharging power factor (default: 1.5)

η_ch::Float64: Charging efficiency (default: 1.0)

η_dch::Float64: Discharging efficiency (default: 1.0)

η_self::Float64: Self-discharge rate (default: 0.0)

α_soc_min::Float64: Minimum state of charge factor (default: 0.0)

α_soc_max::Float64: Maximum state of charge factor (default: 1.0)

lifetime::Int64: Storage lifetime in years (default: 25)

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Lower and upper bounds of storage capacity (default: (lb = 0., ub = 10000.))

Erated_ini::Float64: Initial rated storage energy capacity (default: 1e-6)

soc_ini::Float64: Initial state of charge (default: 0.5)

soh_ini::Float64: Initial state of health (default: 1.0)

Erated::AbstractArray{Float64,2}: Rated storage energy capacity over time and scenarios

carrier::Hydrogen: Hydrogen carrier for the H2 tank

soc::AbstractArray{Float64,3}: State of charge over time, years, and scenarios

cost::AbstractArray{Float64,2}: Cost of the hydrogen storage over time and scenarios

Generation

Genesys.Solar — Type

mutable struct Solar <: AbstractGeneration

A mutable struct representing a solar power generation source. It contains information about the lifetime, bounds, initial conditions, variables, and costs related to the solar generation.

Parameters

-SoH_model::AbstractPVAgingModel (default: FixedLifetimeSolar(25)): The aging model with a fixed lifetime of the solar power generation system. -bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}} (default: (lb = 0., ub = 1000.)): The lower and upper bounds of the solar power generation system. -powerMax_ini::Float64 (default: 0.): The initial maximum power output of the solar power generation system. -soh_ini::Float64 (default: 1.): The initial state of health of the solar power generation system.

Carrier

Genesys.EnergyCarrier — Type

abstract type EnergyCarrier end

An abstract type, parent of all energy carrier types (electricity, heat, hydrogen...)

Genesys.Electricity — Type

Electricity

A mutable struct representing electricity as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with electricity.

Example

Electricity()

Genesys.Heat — Type

Heat

A mutable struct representing heat as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with heat.

Example

Heat()

Genesys.Hydrogen — Type

Hydrogen

A mutable struct representing hydrogen as an energy carrier. It is a subtype of EnergyCarrier.

Fields

power::AbstractArray{Float64,3}: A 3-dimensional array of power values associated with hydrogen.

Hydrogen()

Converter

Genesys.Electrolyzer — Type

Genesys.FuelCell — Type

FuelCell represents the main structure for modeling fuel cell behavior, including efficiency, aging, and other key parameters.

Fields

eff_model::AbstractFuelCellEffModel: The efficiency model used to calculate fuel cell efficiency.

SoH_model::AbstractFuelCellAgingModel: The aging model used to track the State of Health (SoH) of the fuel cell over time.

couplage::Bool: Indicates whether the SoH should influence other parameters (e.g., efficiency, output voltage).

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: Bounds for operational parameters, such as voltage or power.

SoH_threshold::Float64: SoH level at which the fuel cell should be replaced.

V_J_ini::AbstractArray{Float64,2}: Initial voltage-current relationship for the fuel cell.

soh_ini::Float64: Initial SoH at the start of the simulation (default: 1.0, corresponding to 100% health).

N_cell::Int64: Number of cells assembled in the fuel cell stack.

surface::Float64: Surface area of the fuel cell stack (unit should be documented if relevant).

η::AbstractArray{Float64,3}: Efficiency evolution over time (hydrogen to electricity).

carrier::Vector{EnergyCarrier}: Vector of energy carriers used by the fuel cell (e.g., hydrogen, oxygen).

soh::AbstractArray{Float64,3}: Time-evolving State of Health for the fuel cell.

cost::AbstractArray{Float64,2}: Economic metrics associated with fuel cell operation (e.g., cost per hour, maintenance).

Constructor Parameters

eff_model::AbstractFuelCellEffModel=PolarizationFuelCellEfficiency(): Efficiency model to be used.

SoH_model::AbstractFuelCellAgingModel=PowerAgingFuelCell(): Aging model to be used.

couplage::Bool=true: Whether SoH influences other parameters.

bounds=(lb=0.0, ub=50.0): sizing power bounds for all-in-one design.

SoH_threshold::Float64=0.9: SoH replacement threshold (default: 90%).

V_J_ini::AbstractArray{Float64,2}: Initial voltage-current data (default loaded from a CSV file).

soh_ini::Float64=1.0: Initial SoH at the start of the simulation.

Example

fuel_cell = FuelCell(eff_model=PolarizationFuelCellEfficiency(),
                     SoH_model=PowerAgingFuelCell(),
                     couplage=true,
                     bounds=(lb=0.0, ub=60.0),
                     SoH_threshold=0.8,
                     soh_ini=1.0)

This initializes a FuelCell structure with custom sizing power bounds and a lower SoH replacement threshold.

Genesys.Heater — Type

mutable struct Heater <: AbstractConverter

A mutable struct representing a Heater, which is a subtype of AbstractConverter.

Parameters

eff_model::AbstractHeaterEffModel: The model for conversion efficiency of the heater.

SoH_model::AbstractHeaterAgingModel: The aging model for the heater. The default model use a fixed expected lifetime for the heater (in years).

bounds::NamedTuple{(:lb, :ub), Tuple{Float64, Float64}}: The lower and upper bounds of the heater's power.

powerMax_ini::Float64: The initial maximum power output of the heater.

soh_ini::Float64: The initial state of health of the heater.

Variables

powerMax::AbstractArray{Float64,2}: The maximum power output of the heater.

carrier::Vector{Any}: The carrier vector.

cost::AbstractArray{Float64,2}: The cost associated with the heater.

Example

heater = Heater(eff_model=ConstEfficiencyHeater(η_E_H=1.0), SoH_model=FixedLifetimeHeater(lifetime=25), bounds=(lb=30.0, ub=30.0), powerMax_ini=30.0, soh_ini=1.0)

Demand

Genesys.Demand — Type

Demand(; carrier = Electricity())

A mutable struct representing energy demand in the grid. The struct is a subtype of AbstractDemand.

Parameters

carrier::EnergyCarrier: The type of energy carrier for the demand (e.g., Electricity, Heat, etc.) this structre also contains the quantity for each timestamp. see Main.Genesys.EnergyCarrier

timestamp::Array{DateTime,3}: A 3-dimensional array of DateTime objects representing timestamps associated with the demand.

Example

demand = Demand(carrier = Electricity())

grid

Genesys.Grid — Type

mutable struct Grid <: AbstractGrid

A mutable struct representing a Grid, which is a subtype of AbstractGrid.

Parameters

powerMax::Float64: The maximum power capacity of the grid.

Variables

carrier::EnergyCarrier: The energy carrier associated with the grid, such as electricity.

cost_in::AbstractArray{Float64,3}: The input cost associated with the grid.

cost_out::AbstractArray{Float64,3}: The output cost associated with the grid.

cost_exceed::AbstractArray{Float64,2}: The cost associated with exceeding the grid's capacity.

Example

grid = Grid(powerMax=36.0, carrier=Electricity())